Remove binary

+++ shell.c
@@ -14,8 +14,8 @@
 int main(void)
 {
 	char *line = NULL;
-	size_t len = 0;
-	ssize_t nread;
+	unsigned long len = 0;
+	long nread;
 
 	while (1)
 	{
